Removed changelog from index; added disclaimer to page The changelog is incredibly out of date, and does not represent anything recent.  Additionally, the formatting is incorrect as rst (though it could be kept as an md file and continue to use markdown).
It has not been removed, simply is not linked to.  The page now bears disclaimer (which needs to be improved) as to its not-representativeness.

The changelog needs to be improved significantly before it should be displayed.

Additionally, I am unsure of whether a changelog has place in an application's documentation.